[39] 충돌방지 위한 Merge
