+++ src/components/medical/CostInput.tsx
@@ -1,0 +1,23 @@
+import { motion } from "framer-motion";
+import styles from "./CostInput.module.scss";
+
+const CostInput = () => {
+  return (
+    <motion.div
+      className={styles.writeWrapper}
+      initial={{ opacity: 0, x: 100 }}
+      animate={{ opacity: 1, x: 0 }}
+      exit={{ opacity: 0, x: -100 }}
+      transition={{ duration: 0.5, ease: "easeInOut" }}
+    >
+      <label className={styles.writeLabel}>
+        각 치료에 지불한 비용을 <br /> 입력해 주세요.
+      </label>
+      <span className={styles.helperText}>
+        각각의 치료에 해당하는 비용을 입력해 주세요.
+      </span>
+    </motion.div>
+  );
+};
+
+export default CostInput;
