+++ src/styles/_variables.scss
@@ -10,6 +10,7 @@ $white: #fcfcfc;
 $deepblue: #39a7ff;
 $blue: #87c4ff;
 $lightblue: #e0f4ff;
+$blue2: #cde6ff;
 
 // black
